CEST

+++ app/Http/Controllers/ProdutoRestController.php
@@ -67,7 +67,7 @@ class ProdutoRestController extends Controller
 	}
 
 	public function dividePizza(){
-		return response()->json(getenv("DIVISAO_VALOR_PIZZA"), 200);
+		return response()->json((int)getenv("DIVISAO_VALOR_PIZZA"), 200);
 	}
 
 	public function pizzaValorPorTamanho(Request $request){
